+++ csrc/include/flashmoe/context.cuh
@@ -32,7 +32,7 @@ namespace flashmoe
         const auto dispatchTaskQL = cute::ceil_div(world * numLocalExperts,subscriberCount / subscriberWarpSize) *
             (cute::ceil_div(ecTilesM * tilesN0, subscriberWarpSize) + cute::ceil_div(tilesN0, subscriberWarpSize));
         const auto combineTaskQL = cute::ceil_div(ecTilesM * E * tilesN1, subscriberCount);
-        return (dispatchTaskQL + combineTaskQL) * subscriberCount;
+        return static_cast<size_t>(dispatchTaskQL + combineTaskQL) * subscriberCount;
     }
 
     template<int subscriberCount, int subscriberWarpSize>
